extract send-document-email logic from controller
